requirements.txt
flask
flask-sqlalchemy
mysqlclient
qrcode[pil]
pyzbar
pillow
cryptography
python-dotenv
pymysql